@@ -5,7 +5,7 @@ using System.Text.Json;
 
 namespace Runner;
 
-[MemoryDiagnoser, SimpleJob(launchCount: 1, warmupCount: 1, targetCount: 10, invocationCount: 20000)]
+[MemoryDiagnoser, SimpleJob(launchCount: 1, warmupCount: 1, invocationCount: 20000)]
 public class Benchmarks
 {
     private static HttpClient MvcClient { get; } = new WebApplicationFactory<MvcControllers.Program>().CreateClient();
